requirements.txt
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
boto3==1.34.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
django-filter==23.5
geopy==2.4.1
drf-spectacular==0.27.0